+++ lib/modules/manager/npm/extract/common/dependency.ts
